+++ src/full/Agda/TypeChecking/Serialise.hs
@@ -80,7 +80,7 @@ import Agda.Utils.Impossible
 -- 32-bit machines). Word64 does not have these problems.
 
 currentInterfaceVersion :: Word64
-currentInterfaceVersion = 20220625 * 10 + 0
+currentInterfaceVersion = 20220625 * 10 + 1
 
 -- | The result of 'encode' and 'encodeInterface'.